requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
decouple==0.0.7
Django==4.1.6
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
Markdown==3.4.3
mrjob==0.7.4
pipreqs==0.4.11
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarg==0.1.9
